Updates iOS One Signal SDK to 2.12.6. This is a major update since it: - Removes deprecated UIWebView usage from original SDK (https://github.com/OneSignal/OneSignal-iOS-SDK/issues/522) - Fixes a problem while building it with Xcode 11 (https://github.com/williamrijksen/com.williamrijksen.onesignal/issues/85) - Removes deprecated API call to IdsAvailable - Adds CoreGraphics and WebKit must haves frameworks - Updates to Titanium 8.x SDK - Uses TitaniumKit framework and sets CLANG_ENABLE_MODULES to true on project configuration
