+++ ios/Classes/ComWilliamrijksenOnesignalModule.h
@@ -18,7 +18,6 @@ typedef void(^TagsResultHandler)(NSDictionary*, NSError*);
 - (void)deleteTag:(id)args;
 - (void)getTags:(id)value;
 - (void)setLogLevel:(id)args;
-- (void)idsAvailable:(id)args;
 - (void)postNotification:(id)arguments;
 
 @end
