+++ ios/manifest
@@ -2,17 +2,17 @@
 # this is your module manifest and used by Titanium
 # during compilation, packaging, distribution, etc.
 #
-version: 2.1.1
+version: 2.2.0
 apiversion: 2
 architectures: armv7 arm64 i386 x86_64
 description: com.williamrijksen.onesignal
 author: William Rijksen
 license: Specify your license
-copyright: Copyright (c) 2018 by Enrise
+copyright: Copyright (c) 2020 by Enrise
 
 # these should not be edited
 name: com.williamrijksen.onesignal
 moduleid: com.williamrijksen.onesignal
 guid: 67065763-fd5e-4069-a877-6c7fd328f877
 platform: iphone
-minsdk: 7.5.0.GA
+minsdk: 8.0.0.GA
