+++ ios/module.xcconfig
@@ -25,4 +25,4 @@
 //
 // IMPORTANT NOTE: always use $(inherited) in your overrides
 //
-OTHER_LDFLAGS=$(inherited) -framework UserNotifications
+OTHER_LDFLAGS=$(inherited) -framework UserNotifications -framework Webkit -framework CoreGraphics
