+++ ios/platform/OneSignal.framework/Headers
@@ -1,0 +1,1 @@
+Versions/A/Headers
