+++ ios/platform/OneSignal.framework/Modules
@@ -1,0 +1,1 @@
+Versions/A/Modules
