+++ ios/platform/OneSignal.framework/Versions/A/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module OneSignal {
+  umbrella header "OneSignal.h"
+
+  export *
+  module * { export * }
+}
